Limpieza de primavera 2.0
- Reemplazadas clases jerárquicas de Facturae por traits
- Movidas extensiones nativas a `josemmo\Facturae\Extensions`
- Actualizado método `UtilsTrait::getExtension`
