+++ src/FacturaeTraits/ExportableTrait.php
@@ -1,12 +1,12 @@
 <?php
-namespace josemmo\Facturae\Controller;
+namespace josemmo\Facturae\FacturaeTraits;
 
 use josemmo\Facturae\Common\XmlTools;
 
 /**
- * Allows a @link{josemmo\Facturae\Facturae} instance to be exported to XML.
+ * Allows a Facturae instance to be exported to XML.
  */
-abstract class FacturaeExportable extends FacturaeSignable {
+trait ExportableTrait {
 
   /**
    * Add optional fields
@@ -31,9 +31,7 @@ abstract class FacturaeExportable extends FacturaeSignable {
 
   /**
    * Export
-   *
    * Get Facturae XML data
-   *
    * @param  string     $filePath Path to save invoice
    * @return string|int           XML data|Written file bytes
    */
